requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.54.0
httpx==0.27.2
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
